# api/musicAPI.openapi.json
{"openapi":"3.1.0","info":{"title":"musicAPI","description":"","version":"1.0.0"},"tags":[],"paths":{"/songs/delete":{"delete":{"summary":"delete song","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"group":{"type":"string"},"song":{"type":"string"}},"required":["group","song"]}}}},"responses":{"204":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"404":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"500":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/songs/add":{"post":{"summary":"add song","deprecated":false,"description":"","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"group":{"type":"string"},"song":{"type":"string"}},"required":["group","song"]},"example":{"group":"string","song":"string"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"400":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"404":{"description":"doesn`t find in API","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"409":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"500":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}},"examples":{"1":{"summary":"can`t connect to db","value":{"error":"can`t connect to db"}}}}},"headers":{}}},"security":[]}},"/songs/edit":{"patch":{"summary":"edit song","deprecated":false,"description":"Редактирование происходит по всем полям.\nПри отсутствии одного из полей оно изменяется на нулевое.\nДата в формате: DD.MM.YYYY.","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"object","properties":{"group":{"type":"string"},"song":{"type":"string"}},"required":["group","song"]},"info":{"type":"object","properties":{"releaseDate":{"type":"string"},"text":{"type":"string"},"link":{"type":"string"}}}},"required":["title"]},"example":{"title":{"group":"The Beatles","song":"Hey Jude"},"info":{"releaseDate":"12.09.1999","text":"Ooh baby, don't you know I suffer?\\nOoh baby, can you hear me moan?\\nYou caught me under false pretenses\\nHow long before you let me go?\\n\\nOoh\\nYou set my soul alight\\nOoh\\nYou set my soul alight","link":"https://www.example.com/hey-jude"}}}}},"responses":{"204":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"400":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"404":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"500":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}},"examples":{"1":{"summary":"can`t connect to db","value":{"error":"can`t connect to db"}}}}},"headers":{}}},"security":[]}},"/songs/couplet":{"post":{"summary":"get text","deprecated":false,"description":"Куплеты разделяются //n//n, \npage >= 1, limit >= 0.","tags":[],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":0,"schema":{"type":"integer"}},{"name":"limit","in":"query","description":"","required":false,"example":0,"schema":{"type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"group":{"type":"string"},"song":{"type":"string"}},"required":["group","song"]},"example":{"group":"Test","song":"Test"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"couplets":{"type":"string"}},"required":["couplets"]},"examples":{"1":{"summary":"Success","value":{"couplets":"Hey Jude, don't make it bad\nTake a sad song and make it better,\nRemember to let her into your heart\n\nHey Jude, don't make it bad\nTake a sad song and make it better,\nRemember to let her into your heart\n\n"}}}}},"headers":{}},"400":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"404":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"500":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}},"examples":{"1":{"summary":"can`t connect to db","value":{"error":"can`t connect to db"}}}}},"headers":{}}},"security":[]}},"/songs/filter-by-group-and-date":{"post":{"summary":"get songs by group and period","deprecated":false,"description":"Отсутствие массива групп = отсутствие фильтрации по именам группы.\nОтсутствие одной или двух границ периода = отсутствие фильтрации соответственно.\npage >= 1, limit >= 0.","tags":[],"parameters":[{"name":"page","in":"query","description":"","required":false,"schema":{"type":"integer"}},{"name":"limit","in":"query","description":"","required":false,"schema":{"type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"filters":{"type":"object","properties":{"group":{"type":"array","items":{"type":"string"}},"releaseDate":{"type":"object","properties":{"from":{"type":"string"},"to":{"type":"string"}}}},"required":["releaseDate"]}},"required":["filters"]},"example":{"filters":{"group":["Muse"],"releaseDate":{"from":"2006-07-01","to":"2006-07-31"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"title":{"type":"object","properties":{"group":{"type":"string"},"song":{"type":"string"}},"required":["group","song"]},"info":{"type":"object","properties":{"releaseDate":{"type":"string"},"text":{"type":"string"},"link":{"type":"string"}},"required":["releaseDate","text","link"]}},"required":["title","info"]}}},"required":["data"]},"examples":{"1":{"summary":"Example 200","value":{"data":[{"title":{"group":"Muse","song":"Supermassive Black Hole"},"info":{"releaseDate":"16.07.2006","text":"Ooh baby, don't you know I suffer?\\nOoh baby, can you hear me moan?\\nYou caught me under false pretenses\\nHow long before you let me go?\\n\\nOoh\\nYou set my soul alight\\nOoh\\nYou set my soul alight","link":"https://www.youtube.com/watch?v=Xsp3_a-PMTw"}},{"title":{"group":"Muse","song":"Starlight"},"info":{"releaseDate":"27.01.2006","text":"Far away, this ship is taking me...","link":"https://www.youtube.com/watch?v=9jK-NcRmVcw"}},{"title":{"group":"Muse","song":"Time Is on My Side"},"info":{"releaseDate":"22.11.2006","text":"Time is on my side, yes it is...","link":"https://www.youtube.com/watch?v=jQn4PvMZ-1I"}}]}}}}},"headers":{}},"400":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}},"500":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}},"examples":{"1":{"summary":"can`t connect to db","value":{"error":"can`t connect to db"}}}}},"headers":{}}},"security":[]}}},"components":{"schemas":{},"securitySchemes":{}},"servers":[]}